% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol.R
\docType{data}
\name{lol}
\alias{lol}
\title{Measured Predictors from the first 15 minutes of a ranked League of Legends matches.}
\format{
a data frame with 24218 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{MaxTemp}{Daily maximum temperature in degrees Farhenheit.}
}
}
\source{
\url{https://www.kaggle.com/datasets/karlorusovan/league-of-legends-soloq-matches-at-10-minutes-2024/data}
}
\usage{
lol
}
\description{
Measured Predictors from the first 15 minutes of a ranked League of Legends matches.
}
\keyword{datasets}
